third_party/blink/web_tests/virtual/lna-allow-potentially-trustworthy-same-origin/external/wpt/fetch/local-network-access/redirect.https.window-expected.txt

This is a testharness.js-based test.
PASS local to private to public: success.
PASS local to private to local: success.
PASS private to private to local: failed preflight.
PASS private to private to local: success.
PASS private to private to local: no-cors success.
PASS private to local to private: failed preflight.
PASS private to local to private: success.
PASS private to local to private: no-cors success.
PASS public to private to local: failed first preflight.
PASS public to private to local: failed second preflight.
PASS public to private to local: success.
PASS public to private to local: no-cors success.
FAIL treat-as-public to local to private: failed first preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL treat-as-public to local to private: failed second preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL treat-as-public to local to private: success. assert_equals: response ok expected true but got false
FAIL treat-as-public to local to private: no-cors failed first preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL treat-as-public to local to private: no-cors failed second preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL treat-as-public to local to private: no-cors success. assert_equals: response body expected "" but got "no preflight received"
PASS treat-as-public to private to local: failed first preflight.
FAIL treat-as-public to private to local: failed second preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL treat-as-public to private to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Failed to fetch"
PASS treat-as-public to private to local: no-cors failed first preflight.
FAIL treat-as-public to private to local: no-cors failed second preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS treat-as-public to private to local: no-cors success.
Harness: the test ran to completion.
